+++ api/task/__init__.py
@@ -1,3 +1,4 @@
+import sys
 import os
 import io
 import logging
@@ -11,10 +12,20 @@ from ohdsi.database_connector import (
     disconnect,
 )
 
+log_stream = io.StringIO()
+
+logging.basicConfig(
+    level=logging.DEBUG,
+    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+    handlers=[logging.StreamHandler(sys.stdout), logging.StreamHandler()],
+)
+
 log = logging.getLogger(__name__)
+log.setLevel(logging.DEBUG)
 
 
 class OHDSITask(Task):
+
     def __init__(self):
         super().__init__()
 
@@ -25,18 +36,17 @@ class OHDSITask(Task):
         self.password = os.environ.get("OMOP_PASSWORD")
         self.port = os.environ.get("OMOP_PORT")
         self.schema = os.environ.get("OMOP_CDM_SCHEMA")
+        self.result_schema = os.environ.get("OMOP_RESULT_SCHEMA")
 
-        self.log_stream = io.StringIO()
         self.connection = None
 
         signals.task_prerun.connect(self.on_task_prerun)
-        # signals.task_success.connect(self.on_task_success)
+        signals.task_success.connect(self.on_task_success)
 
     def on_task_prerun(self, task_id, task, *args, **kwargs):
         # self.update_state(state='PROGRESS', meta={'current': 2, 'total': 10})
-        log.info("Creating connection details")
-
-        self.update_state(state="INITIALIZE", meta={"step": 1, "steps": 5})
+        log.debug("Creating connection details")
+        self.update_state(state="INITIALIZE")
         connection_details = create_connection_details(
             self.dbms,
             server=f"{self.server}/{self.database}",
@@ -45,13 +55,22 @@ class OHDSITask(Task):
             port=self.port,
         )
 
-        log.info("Connecting to database")
-        self.update_state(state="CONNECT", meta={"step": 2, "steps": 5})
-        with redirect_stdout(self.log_stream):
-            self.connection = connect(connection_details)
+        log.debug("Connecting to database")
+        task.update_state(state="CONNECTING")
+        self.connection = connect(connection_details)
+
+        task.request.update(
+            {
+                "schema": self.schema,
+                "result_schema": self.result_schema,
+                "connection": self.connection,
+                "dbms": self.dbms,
+            }
+        )
+        task.update_state(state="EXECUTING")
 
     def on_task_success(self, result, **kwargs):
-        log.info("Disconnecting from database")
-        # TODO: this part is not returned to the client
-        # with redirect_stdout(self.log_stream):
+        log.debug("Disconnecting from database")
+        # TODO: return the logstream to the user
+        print(log_stream.getvalue())
         disconnect(self.connection)