+++ api/task/count.py
@@ -11,15 +11,18 @@ log = logging.getLogger(__name__)
 
 @shared_task(bind=True, ignore_result=False, time_limit=60, base=OHDSITask)
 def count(self):
-    print("Going to count the number of persons in the database")
-    if not self.schema:
+
+    log.debug("Going to count the number of persons in the database")
+    if not self.request.schema:
         return {"error": "No schema is set"}
 
-    count = query_sql(self.connection, f"SELECT COUNT(*) FROM {self.schema}.person")
-    print("Counted the number of persons in the database")
+    count = query_sql(
+        self.request.connection, f"SELECT COUNT(*) FROM {self.request.schema}.person"
+    )
+    log.debug("Counted the number of persons in the database")
 
     count = RS4Extended.from_RS4(count)
     count = convert_from_r(count.extract("COUNT"))
-    print("Converted from R object")
+    log.debug("Converted from R object")
 
     return count